requirements.txt
# Core Dependencies
python-dotenv==1.0.0
google-generativeai==0.3.2

# PDF Processing
PyPDF2==3.0.1

# Text Processing & Embeddings
langchain==0.1.20
sentence-transformers==2.5.1

# Vector Database
chromadb==0.4.22

# Web Interface
streamlit==1.32.0

# Utilities
numpy>=1.26.0
tqdm==4.66.1